+++ lib/QueryableSelector.php
@@ -4,13 +4,13 @@ declare(strict_types=1);
 
 namespace PeachySQL;
 
-use PeachySQL\QueryBuilder\Selector;
+use PeachySQL\QueryBuilder\{Selector, SqlParams};
 
 class QueryableSelector extends Selector
 {
     private PeachySql $peachySql;
 
-    public function __construct(string $query, PeachySql $peachySql)
+    public function __construct(SqlParams $query, PeachySql $peachySql)
     {
         parent::__construct($query, $peachySql->getOptions());
         $this->peachySql = $peachySql;